requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
joblib
werkzeug
requests
python-dotenv